requirements.txt
# ----------------------
# Flask
# ----------------------
click==6.7
flask
Flask-WTF==0.14.2
itsdangerous==0.24
Flask-SQLAlchemy
Flask-Migrate
psycopg2
Flask-Script
pyjwt
Werkzeug==0.11.15
WTForms==2.1
Jinja2==2.9.4
# ----------------------
# Production Server
# ----------------------
gunicorn==19.6.0
